Created nav bar and CSS theme styling
A nav bar was created to cycle between the three required pages (D&D, MTG and Settings). Icons were added and sized accordingly.

Tailwind styling was configured for various parameters used when styling this bar, which includes colours for light and dark mode. App.tsx and the use of the 'dark' class allows for automatic and manual toggling of light/dark mode, which was applied to index.tsx and the nav bar.